+++ src/main/java/byuntil/backend/research/domain/entity/Project.java
@@ -1,6 +1,6 @@
 package byuntil.backend.research.domain.entity;
 
-import byuntil.backend.research.dto.ProjectDto;
+import byuntil.backend.research.dto.request.ProjectDto;
 import lombok.*;
 
 import javax.persistence.*;
