+++ src/main/java/byuntil/backend/research/service/ThesisService.java
@@ -5,7 +5,7 @@ import byuntil.backend.research.domain.entity.Field;
 import byuntil.backend.research.domain.entity.Thesis;
 import byuntil.backend.research.domain.repository.FieldRepository;
 import byuntil.backend.research.domain.repository.ThesisRepository;
-import byuntil.backend.research.dto.ThesisDto;
+import byuntil.backend.research.dto.request.ThesisDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
