+++ src/test/java/byuntil/backend/common/IntegrationTest.java
@@ -1,19 +1,15 @@
 package byuntil.backend.common;
 
 import byuntil.backend.member.domain.entity.member.Member;
-import byuntil.backend.member.domain.repository.MemberRepository;
 import byuntil.backend.member.dto.request.ProfessorSaveRequestDto;
 import byuntil.backend.member.service.MemberService;
 import byuntil.backend.member_thesis.entity.Member_Thesis;
 import byuntil.backend.member_thesis.service.Member_ThesisService;
-import byuntil.backend.post.domain.entity.Board;
-import byuntil.backend.post.domain.entity.Post;
 import byuntil.backend.post.domain.repository.BoardRepository;
 import byuntil.backend.post.dto.PostDto;
 import byuntil.backend.research.domain.entity.Thesis;
-import byuntil.backend.research.dto.FieldDto;
-import byuntil.backend.research.dto.ProjectDto;
-import byuntil.backend.research.dto.ThesisDto;
+import byuntil.backend.research.dto.request.FieldDto;
+import byuntil.backend.research.dto.request.ThesisDto;
 import byuntil.backend.research.service.ThesisService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -39,7 +35,7 @@ public class IntegrationTest {
     @Transactional
     @Commit
     //member 여러명 thesis 여러명 해서 다대다 test(thesis, member, field)
-    public void thesis_member_field(){
+    public void thesis_member_field() {
         //given
         ProfessorSaveRequestDto dto1 = makeProfessorDto();
         ProfessorSaveRequestDto dto2 = makeProfessorDto();
@@ -48,10 +44,10 @@ public class IntegrationTest {
         FieldDto fieldDto1 = makeFieldDto();
         //when
         Long memberId = memberService.saveMember(dto1);
-        Member member = (Member)memberService.findOneMember(memberId).get();
+        Member member = (Member) memberService.findOneMember(memberId).get();
 
         Long memberId2 = memberService.saveMember(dto2);
-        Member member2 = (Member)memberService.findOneMember(memberId2).get();
+        Member member2 = (Member) memberService.findOneMember(memberId2).get();
 
         //field를 thesis에 넣기
         Long thesisId1 = thesisService.save(thesisDto1);
@@ -64,8 +60,8 @@ public class IntegrationTest {
         Long MTId1 = member_thesisService.createThesis(memberId, thesisId1);
         Long MTId2 = member_thesisService.createThesis(memberId2, thesisId2);
 
-        Member_Thesis memberThesis1 =member_thesisService.findById(MTId1).get();
-        Member_Thesis memberThesis2 =member_thesisService.findById(MTId2).get();
+        Member_Thesis memberThesis1 = member_thesisService.findById(MTId1).get();
+        Member_Thesis memberThesis2 = member_thesisService.findById(MTId2).get();
         //then
 
         //1. 중간테이블로 매핑이 잘되었는가?
@@ -75,29 +71,33 @@ public class IntegrationTest {
         //3. field가 안넣어졌는가?
         Assertions.assertThat(memberThesis2.getThesis().getField()).isNull();
     }
+
     //attach를 포함한 post를 member가 올리기
     @Transactional
     @Commit
     @Test
-    public void attach_post_board_member(){
+    public void attach_post_board_member() {
         //given
         //board는 이미 추가해놓은 상태임
         PostDto postDto = makePostDto();
         //when
         //then
     }
-    public PostDto makePostDto(){
+
+    public PostDto makePostDto() {
         return PostDto.builder().title("제목1").boardName("board1").content("내용11").build();
     }
 
-    public FieldDto makeFieldDto(){
-        return FieldDto.builder().name("field").description(""+Math.random()).build();
+    public FieldDto makeFieldDto() {
+        return FieldDto.builder().name("field").description("" + Math.random()).build();
     }
-    public ThesisDto makeThesisDto(){
-        return ThesisDto.builder().enName(""+Math.random()).koName("한글이름").journal("sss")
+
+    public ThesisDto makeThesisDto() {
+        return ThesisDto.builder().enName("" + Math.random()).koName("한글이름").journal("sss")
                 .title("ss").fieldName("field1").publishDate(LocalDateTime.now()).url("sda/ss").build();
     }
-    public ProfessorSaveRequestDto makeProfessorDto(){
+
+    public ProfessorSaveRequestDto makeProfessorDto() {
 
         ProfessorSaveRequestDto professor = ProfessorSaveRequestDto.builder()
                 .email("asdfa")
