+++ src/test/java/byuntil/backend/research/ResearchTest.java
@@ -7,9 +7,9 @@ import byuntil.backend.research.domain.repository.DemoRepository;
 import byuntil.backend.research.domain.repository.FieldRepository;
 import byuntil.backend.research.domain.repository.ProjectRepository;
 import byuntil.backend.research.domain.repository.ThesisRepository;
-import byuntil.backend.research.dto.DemoDto;
-import byuntil.backend.research.dto.FieldDto;
-import byuntil.backend.research.dto.ProjectDto;
+import byuntil.backend.research.dto.request.DemoDto;
+import byuntil.backend.research.dto.request.FieldDto;
+import byuntil.backend.research.dto.request.ProjectDto;
 import byuntil.backend.research.service.DemoService;
 import byuntil.backend.research.service.FieldService;
 import byuntil.backend.research.service.ProjectService;
