+++ src/test/java/byuntil/backend/research/controller/common/ProjectControllerTest.java
@@ -2,8 +2,8 @@ package byuntil.backend.research.controller.common;
 
 import byuntil.backend.research.domain.entity.Field;
 import byuntil.backend.research.domain.entity.Project;
-import byuntil.backend.research.dto.FieldDto;
-import byuntil.backend.research.dto.ProjectDto;
+import byuntil.backend.research.dto.request.FieldDto;
+import byuntil.backend.research.dto.request.ProjectDto;
 import byuntil.backend.research.service.FieldService;
 import byuntil.backend.research.service.ProjectService;
 import org.junit.jupiter.api.DisplayName;
